Fix #380 to add backward compatibility to the NewVersionForm, allowing non-zip download urls.

+++ XrmToolBox/Forms/NewVersionForm.cs
@@ -39,8 +39,9 @@ namespace XrmToolBox.Forms
         {
             var currentAssemblyFolder = new FileInfo(Assembly.GetExecutingAssembly().FullName).DirectoryName;
             var updaterFile = Path.Combine(currentAssemblyFolder, "XrmToolBox.AutoUpdater.exe");
+            var urlIsZip = downloadUrl.ToString().ToLowerInvariant().EndsWith(".zip");
 
-            if (File.Exists(updaterFile))
+            if (File.Exists(updaterFile) && urlIsZip)
             {
                 var destinationFolder =
                     Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.CommonApplicationData),
@@ -60,8 +61,11 @@ namespace XrmToolBox.Forms
             }
             else
             {
-                MessageBox.Show(this,
-                    "Auto updater has not been found! The new version will be downloaded only. Please install it manually");
+                if (urlIsZip)
+                {
+                    MessageBox.Show(this,
+                        "Auto updater has not been found! The new version will be downloaded only. Please install it manually");
+                }
                 Process.Start(downloadUrl.ToString());
             }
         }
